Move various file to a git status dub directory
